Fully replace old table widget with refactored version
Remove all code related to the old table widget and replace it with the revised version.

WIDGETS FRNT-2 [IN PROGRESS]
Table FRNT-3 [IN PRGORESS]
Table widget refactor FRNT-100 [IN PROGRESS] -> [DONE]
