+++ BaCa2/templates/admin.html
@@ -14,7 +14,7 @@
     </div>
     <div class="tab-wrapper" id="courses-table-tab">
         <div class="side-nav-tab">
-            {% with widgets.TableWidget.courses_table as table_widget %}
+            {% with widgets.TableWidget.course_table_widget as table_widget %}
                 {% include "widget_templates/table.html" %}
             {% endwith %}
         </div>
