+++ BaCa2/templates/courses.html
@@ -1,5 +1,4 @@
 {% extends 'base.html' %}
 
 {% block content %}
-    {% include 'widget_templates/table.html' %}
 {% endblock %}
