+++ BaCa2/templates/dashboard.html
@@ -2,8 +2,8 @@
 
 {% block content %}
     <div class="row">
-            {% with widgets.TableWidget2.course_table_widget as table_widget %}
-                {% include 'widget_templates/table2.html' %}
+            {% with widgets.TableWidget.course_table_widget as table_widget %}
+                {% include 'widget_templates/table.html' %}
             {% endwith %}
     </div>
