--- BaCa2/templates/widget_templates/table2.html
@@ -1,39 +1,0 @@
-{% load custom_tags %}
-
-{% if table_widget.allow_delete %}
-{% endif %}
-
-<div class="table-wrapper card">
-    <div class="card-body">
-        <div class="table-responsive">
-            <table class="{{ table_widget.table_class }}"
-                   id="{{ table_widget.name }}"
-                   style="width:100%">
-                <thead>
-                    <tr>
-                        {% for col in table_widget.cols %}
-                            <th class="{{ col.col_type }}">
-                                {{ col.header }}
-                            </th>
-                        {% endfor %}
-                    </tr>
-                </thead>
-            </table>
-        </div>
-    </div>
-</div>
-
-
-<script type="text/javascript">
-    initTable({
-        tableId: '{{ table_widget.name }}',
-        modelName: '{{ table_widget.model_name }}',
-        cols: {{ table_widget.cols|safe }},
-        defaultOrder: '{{ table_widget.default_order }}',
-        defaultOrderCol: parseInt({{ table_widget.default_order_col }}),
-        paging: {{ table_widget.paging|safe }},
-        refresh: {{ table_widget.refresh|safe }},
-        refreshInterval: parseInt({{ table_widget.refresh_interval }}),
-    });
-</script>
-
